refactor: 약자로 되어있던 패키지 이름 수정 - req -> request - res -> response
