+++ src/main/java/com/moa/user/vo/response/UserSignUpResponse.java
@@ -1,4 +1,4 @@
-package com.moa.user.vo.res;
+package com.moa.user.vo.response;
 
 
 import lombok.Getter;
